:truck: Refactor : make utils dir #56
