+++ backend/routes/lentRouter.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const lentRouter = express.Router();
 const { getLentUserInfo, getLentOverdue } = require('../db/query');
-const { sendResponse } = require('../util');
+const { sendResponse } = require('../utils/util');
 const pool = require('../config/database');
 
 const getLentInfo = async (_req, res) => {
@@ -10,9 +10,6 @@ const getLentInfo = async (_req, res) => {
   try {
     const lentInfo = await getLentUserInfo(connection);
     return sendResponse(res, lentInfo, 200);
-  } catch (err) {
-    console.log(err);
-    throw err;
   } finally {
     connection.release();
   }
@@ -23,9 +20,6 @@ const getLentOverdueInfo = async (_req, res) => {
   try {
     const overdueInfo = await getLentOverdue(connection);
     return sendResponse(res, overdueInfo, 200);
-  } catch (err) {
-    console.log(err);
-    throw err;
   } finally {
     connection.release();
   }
